backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bidict==0.23.1
blinker==1.9.0
blis==1.3.3
catalogue==2.0.10
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.13
dataclasses-json==0.6.7
distro==1.9.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
frozenlist==1.8.0
greenlet==3.0.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.7
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
# langchain-experimental removed for lite mode
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.6
lxml==5.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
multidict==6.7.1
murmurhash==1.0.15
mypy_extensions==1.1.0
networkx==3.2.1
numpy==1.26.4
openai==2.16.0
orjson==3.11.5
ormsgpack==1.12.2
packaging==25.0
pandas==2.1.4
playwright==1.40.0
pluggy==1.6.0
preshed==3.0.12
propcache==0.4.1
psycopg2-binary==2.9.9
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyee==11.0.1
Pygments==2.19.2
pytest==7.4.4
pytest-asyncio==0.23.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
python-engineio==4.13.0
python-socketio==5.16.0
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.1
setuptools==80.10.2
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
# spacy removed for lite mode
SQLAlchemy==2.0.46
srsly==2.5.2
tenacity==9.1.2
# thinc removed for lite mode
tiktoken==0.12.0
tqdm==4.67.1
typer==0.12.5
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.0
wasabi==1.1.3
weasel==0.4.3
Werkzeug==3.1.5
wrapt==2.0.1
wsproto==1.3.2
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
pymupdf==1.25.3
rapidfuzz==3.12.1
